refactor: room.ts를 item/index.tx로 옮김
