+++ src/conversation-events/interfaces/socket-information.interface.ts
@@ -1,4 +1,4 @@
-import { logger } from 'src/rooms/room';
+import { logger } from 'src/rooms/item';
 import { RoomSocket, SocketStatus } from './room-socket.interface';
 
 //TODO: 여기 다른 것 추가해야하는지 front에게 묻기
